Add fallback media layout for notification actions
Adds more button slots which are dynamically shown depending on which actions
are supported for the app's media.

When PlaybackState actions are available, the large play/pause button
will appear, along with prev/next on either side of the progress bar.
When Notification actions are used, those buttons will be hidden and up
to 5 actions will be shown to the right of the progress bar instead.

Also refactored view holder classes and test since both templates have 5
generic action buttons now.

Fixes: 218875984
Bug: 218876469
Test: atest MediaControlPanelTest
Change-Id: I464834d2eee53d661f2407f6280e2856c96c63ac
Signed-off-by: Mr_ZerO <rbesar210311@gmail.com>